Refactor LocationList: Extract LocationRow to separate component
